Renamed convenience_wrappers.py -> differential_operators.py
