+++ autograd/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import absolute_import
-from .convenience_wrappers import (
+from .differential_operators import (
     make_vjp, grad, multigrad_dict, elementwise_grad, value_and_grad,
     grad_and_aux, hessian_tensor_product, hessian_vector_product, hessian,
     jacobian, tensor_jacobian_product, vector_jacobian_product, grad_named,
